requirements.txt
## The following requirements were added by pip freeze:
alembic==1.8.1
aniso8601==9.0.1
blinker==1.5
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-HTTPAuth==4.7.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
onetimepass==1.0.1
Pillow==9.3.0
PyJWT==2.6.0
PyQRCode==1.2.1
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.44
text-unidecode==1.3
Werkzeug==2.2.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.10.0

